+++ src/Completion/Message/AssistantMessage.php
@@ -5,16 +5,14 @@ declare(strict_types=1);
 namespace ByCerfrance\LlmApiLib\Completion\Message;
 
 use ByCerfrance\LlmApiLib\Completion\Content\ContentInterface;
-use ByCerfrance\LlmApiLib\Completion\Content\TextContent;
 use ByCerfrance\LlmApiLib\Completion\Tool\ToolCall;
 use Override;
 
 /**
  * Assistant message with optional tool calls.
  */
-readonly class AssistantMessage implements MessageInterface
+readonly class AssistantMessage extends Message
 {
-    private ContentInterface $content;
     /** @var ToolCall[] */
     private array $toolCalls;
 
@@ -26,30 +24,12 @@ readonly class AssistantMessage implements MessageInterface
         string|ContentInterface|null $content = null,
         array $toolCalls = [],
     ) {
-        if (is_string($content)) {
-            $this->content = new TextContent($content);
-        } elseif (null === $content) {
-            $this->content = new TextContent('');
-        } else {
-            $this->content = $content;
-        }
+        parent::__construct($content, RoleEnum::ASSISTANT);
 
         $this->toolCalls = array_filter(
             $toolCalls,
             fn($v) => $v instanceof ToolCall,
         );
-    }
-
-    #[Override]
-    public function getRole(): RoleEnum
-    {
-        return RoleEnum::ASSISTANT;
-    }
-
-    #[Override]
-    public function getContent(): ContentInterface
-    {
-        return $this->content;
     }
 
     /**
@@ -75,23 +55,13 @@ readonly class AssistantMessage implements MessageInterface
     #[Override]
     public function jsonSerialize(): array
     {
-        $data = [
-            'role' => $this->getRole(),
-        ];
-
-        if (!$this->hasToolCalls()) {
-            $data['content'] = $this->content;
-        } else {
-            $data['content'] = null;
-            $data['tool_calls'] = $this->toolCalls;
+        if ($this->hasToolCalls()) {
+            return [
+                'role' => $this->getRole(),
+                'tool_calls' => $this->getToolCalls(),
+            ];
         }
 
-        return $data;
-    }
-
-    #[Override]
-    public function requiredCapabilities(): array
-    {
-        return $this->content->requiredCapabilities();
+        return parent::jsonSerialize();
     }
 }